+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/captcha/MapCaptchaInfo.java
@@ -33,13 +33,13 @@ public final class MapCaptchaInfo {
   private final FallbackPacket[] legacy;
   private final FallbackPacket modern;
 
-  public MapCaptchaInfo(final @NotNull String answer, final byte @NotNull [] buffer) {
+  public MapCaptchaInfo(final @NotNull String answer, final int @NotNull [] buffer) {
     this.answer = answer;
 
     // Prepare 1.7 map data using a grid
-    final byte[][] grid = new byte[128][128];
+    final int[][] grid = new int[128][128];
     for (int i = 0; i < buffer.length; i++) {
-      final byte buf = buffer[i];
+      final int buf = buffer[i];
       grid[i & Byte.MAX_VALUE][i >> 7] = buf;
     }
     this.legacy = new FallbackPacket[grid.length];
